created a sepearte component for comments section.
